requirements.txt
beautifulsoup4==4.12.3
black==24.8.0
cachetools==5.5.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
coverage==7.6.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
django==4.2
django-audit-fields==0.3.17
django-auditlog==3.0.0
django-defender==0.9.8
django-environ==0.11.2
django-import-export==4.1.1
django-revision==0.3.7
django-simple-history==3.7.0
django-webtest==1.9.12
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
drf-yasg==1.21.7
edc-device==0.3.17
edc-utils==0.3.34
filelock==3.15.4
flake8==7.1.1
gitdb==4.0.11
GitPython==3.1.43
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
model-bakery==1.19.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pyproject-api==1.7.1
pytest==8.3.2
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
tablib==3.5.0
tox==4.18.0
uritemplate==4.1.1
virtualenv==20.26.3
waitress==3.0.0
WebOb==1.8.8
WebTest==3.0.1
wheel==0.43.0
xmltodict==0.13.0